Add a test for fullscreen for the A/B/A case
See the discussion at [1] for more context. This test requests
fullscreen on A, then B, then A, where A and B are sibling divs. All
three requests should succeed, and at the end, A should be topmost (and
the fullscreen element) with both A and B in the top layer.

[1] https://github.com/whatwg/fullscreen/pull/223

Change-Id: I3f35dda5b9eb1bc24201616bb5bb4949d20fd170