[A11y] Enhance performance by targeting value updates to specific nodes
This is a performance improvement mostly for ProcessDeferredAccessibilityEvents(), although it does seem to improve SendPendingAccessibilityEvents() and in some cases unserialization in BrowserAccessibilityManager::OnAccessibilityEvents() as well. Also fixes some crashes in referenced issues.

Previously, when a change occurs to the DOM or layout tree, accessibility increments AXObjectCacheImpl::modification_count_, which has the affect of invalidating the cached values on all AXObjects in the tree. The next time one of those properties is fetched, AXObject::UpdateCachedAttributeValuesIfNeeded() will recompute all of the cached values for that AXObject.

However, for most changes, it is only necessary to update the cached values for the AXObject that has changed, or potentially for a subtree. This CL implements that by providing 3 methods:
* AXObject::InvalidateCachedValues() for invalidating a single AXObject
* InvalidateCachedValuesOnSubtreeWithFlatTraversal(Node*), which is useful for visibility:hidden and aria-hidden toggles
* InvalidateCachedValuesOnAllObjects() for invalidating all AXObjects in the tree.

Explanation of test changes:
* content-visibility tests: redundant whitespace nodes are now removed once a subtree gains layout, so that the AXObject subtree looks identical whether the layout was provided from the beginning or gained later
* visibility:hidden tests: added more content because we lacked test coverage of deep subtree invalidations.
* Linux event tests: redundant events are now removed
* other-aria-attribute-change-sends-notification.html: AriaAttributeChanged is no longer needed
* is-ignored-change-sends-notification.html: improvement as redundant notification is no longer fired, because the visibility hidden child does not change it's tree inclusion status (visibility hidden objects are currently ignored but included in the tree).

Follow-ups:
* Try to remove Event::kAriaAttributeChanged everywhere, it is not necessary
* Look into queuing up fewer callbacks for attribute changes

Fixed: 1434555, 1362758
Cq-Do-Not-Cancel-Tryjobs: true
Change-Id: Iecb56b877392954e7ce1fea37a9472b68ba5364f