Change web expose navigation id to random UUID string
This CL changes the navigation id of the LocalDOMWindow from an int to a
random UUID that is a string. The nav id increments each time there is a
soft nav or bfcache nav. Now the nav id would change to another random
UUID. This is to make the navigation id non-sequential so we can record
it into UKM.

Things to note
1, The WTF::CreateCanonicalUUIDString() is used, instead of the
Crypto::randomUUID(). These 2 are the same. The latter is used only in
v8 code and the former is used in all other places.

2, Previously the navigation is an int with a default number. So if the
nav id is the default number, we know no soft nav or bfcache nav happens.
With nav ids being UUIDs now, the nav id at the LocalDOMWindow creation
time is initiated to a UUID no different than other UUIDs generated after
each soft nav or bfcache nav. Thus by just looking UUIDs, we no longer
know if the current Dom has gone thru soft nav and bfcache nav unless we
keep the initial UUID and compare. It is not a problem. Just note the
difference.

3, The test case, PerformanceTimelineNavigationIdBrowserTest
_NonBackForwardCacheRestore is removed as now the nav id would be
re-generated and different regardless navigation type.
The other test case PerformanceTimelineNavigationIdBrowserTest
_BackForwardCacheRestore is modified accordingly. It now verifies that
each bfcache navigation would change the nav id too.

Bug:1370937
Change-Id: Ifb18df3f267ac39e98d868e558758f400222bd43