Pre-paint: Set up properties for missed OOF descendants.
Set up paint properties as correctly as possible for ancestors of a
missed out-of-flow positioned descendant. OOF fragments may be missed
when block-fragmented, when some ancestor between the OOF itself and the
fragmentainer isn't represented in that fragmentainer. Previously we'd
just miss out on all paint properties specified on such ancestors (e.g.
opacity, transforms, clipping).

When setting up paint properties in this mode, side-effects will be
disabled for the ancestors that we're processing - i.e. we'll leave
LayoutObject, PaintLayer and FragmentData alone. We'll just set up the
temporary context objects for pre-paint correctly, so that the missed
descendants get *their* FragmentData objects set up correctly, with
transforms, opacity, etc.

Note that, for now we're skipping this for all clip-related properties.
We need to fix crbug.com/1442211 first, or we risk clipping content that
really should be visible. For other properties like transforms this is
less of an issue. Prior to this change, transforms wouldn't be applied
at all, while, with this CL, they will, although with a potentially
incorrect transform origin. For other properties again, like opacity or
filters, this shortcoming is of no concern at all, since they don't
depend on the offset or size of the fragment (which we currently get
wrong in some cases, due to the bug mentioned).

Added a run-time enabled feature kill-switch, just in case.

Bug: 1371426
Change-Id: I2fcebaba913907c09b7cd182053aae7fc8689179