Inherit COOP: restrict-properties and un-suppress opener
To make the `Cross-Origin-Opener-Policy: restrict-properties` (COOP:RP)
more useful, we'd like to let cross-origin iframes inherit this policy
from their parent and keep the opener connection between windows that
the iframes open. To make sure new windows are in correct browsing
instances, we need to keep track of the origin that sets the policy when
running COOP algorithms. See
https://github.com/hemeryar/coi-with-popups/blob/main/docs/cross_origin_iframe_popup.MD
for more explanations.

In https://crrev.com/c/4404116, we've bundled the origin that sets the
COOP with the policy.  This CL is to use the stored origin in COOP
algorithms:
1. In `RenderFrameHostImpl::DidCommitNewDocument`, we let embedded
documents inherit the COOP:RP bundle when the policy is COOP:RP.
2. To keep the opener connection, we return false in
`COOPSuppressOpener` when COOP value is RP.
3. Since we allow the popup's initial empty document to stay in the
process, we had to restrict its crossOriginIsolated capability.
Otherwise, a cross-origin iframe can get around permission policy that
restricts it from using crossOriginIsolated APIs. Therefore, we put the
COOP bundle into the policy container so it is passed down to the
renderer. We can then decide whether we need to restrict
crossOriginIsolated.

Bug: 1385827
Change-Id: I39cd140f70b628114205640af2220ec2de45ac06