Improve hover triggering functionality to support popover=manual
This CL primarily adds a much larger hover hiding test, which tests
all popover types, two invoker types, and all popovertargetaction
values. In building that, I found a number of cases that weren't
properly handled, mostly because hover detection was using a version
of IsNodePopoverDescendant that was fairly manual and old. This
new version makes better use of a newly-expanded
FindTopmostRelatedPopover which is able to optionally look at manual
popovers. In doing this, I also refactored the `inclusive` argument
into an EnumSet to allow controlling the manual popover behavior
at the same time.

Bug: 1416284
Change-Id: I2370610b4c7517fa635c353f0cfaf6de7b16b3af